README.md
Research:
The bill is a document that is sent to the customers from the seller or service provider about the
money customers owe to the buyer. It consists of various information about the services or
goods provided by the seller or service provider. The electricity bill contains information like
customer ID, customer name, customer telephone number, customer type
(domestic/commercial) and the number of units consumed. It should also be in a proper manner
so that the customer can read it properly without any problem.
Analyse:
Requirements:
1. The customer ID, customer name, customer telephone number, customer type
(domestic/commercial) , and number of units are needed .
2. Telephone numbers should be only 10 digits.
3. The price of the units consumed should be correct.
4. The period of consumption should be mentioned in the proper format.
5. The bill should be provided in proper format that should include the information
mentioned above.
Challenges:
1. The telephone number that is entered can be more or less than 10 digits.
2. The prices of the units consumed differ for both the types of customers so it should be
mentioned.
3. Customer ID is an important aspect in the process as it contains the information about
the customers so it should be correct.
Ideate:
To design a software to generate the bill of the hotel that should include the customer
information in an aesthetic manner to satisfy the customers. The bill should include the
information such as:
1. Customer name
2. Customer telephone number
3. Types of rooms booked
4. Number of rooms booked
5. Room number
6. Valid I'd Proof
It should also include the information of the hotel such as name, contact number, registration
number and address. In the end , it should also include a thanking message to customers for
using their hotel.
Test:
This program simulates a hotel billing system.
It:
•Takes customer details (name, Aadhaar, phone)
•Lets the customer choose room type and count
•Calculates bill including GST and service charge
•Prints a formatted bill/receipt